buffer pairs are in buffer_pair.hpp
