+++ test/unit/CMakeLists.txt
@@ -20,14 +20,13 @@ set(PFILES
     any_dynamic_buffer.cpp
     asio.cpp
     buffer.cpp
+    buffer_pair.cpp
     buffers.cpp
     circular_buffer.cpp
-    const_buffer_pair.cpp
     copy.cpp
     flat_buffer.cpp
     front.cpp
     make_buffer.cpp
-    mutable_buffer_pair.cpp
     range.cpp
     slice.cpp
     string_buffer.cpp
