+++ test/unit/front.cpp
@@ -13,7 +13,7 @@
 #if 0
 
 #include <boost/buffers/copy.hpp>
-#include <boost/buffers/const_buffer_pair.hpp>
+#include <boost/buffers/buffer_pair.hpp>
 #include <boost/buffers/prefix.hpp>
 #include <boost/buffers/sans_prefix.hpp>
 #include <boost/buffers/sans_suffix.hpp>
